+++ src/diagnostics/diagnostics.ts
@@ -1,8 +1,13 @@
 import * as vscode from 'vscode';
 
+function crushControlChars(lineText: string): string {
+    const crushed = lineText.replace(/\{(\w+|-|\s|\d+|\/|\.)*\}/gi, "_");
+    return crushed;
+}
+
 const lineChecks = [
     {
-        check:(lineOfText: vscode.TextLine) => lineOfText.text.length > 80,
+        check:(lineOfText: vscode.TextLine) => crushControlChars(lineOfText.text).length > 80,
         createDiagnostic: (doc: vscode.TextDocument, lineOfText: vscode.TextLine, lineIndex: number): vscode.Diagnostic => {
             const index = 0;
             const range = new vscode.Range(lineIndex, index, lineIndex, lineOfText.text.length);
@@ -13,9 +18,11 @@ const lineChecks = [
     }
 ];
 
-export function updateDiagnostics(document: vscode.TextDocument, collection: vscode.DiagnosticCollection): void {
-	if (document) {
-        const diagnostics: vscode.Diagnostic[] = [];
+export function checkDocument(document: vscode.TextDocument): vscode.Diagnostic[] {
+    const configuration = vscode.workspace.getConfiguration('c64basicv2');
+    const enableDiagnostics : boolean | undefined = configuration.get("enableDiagnostics");
+    const diagnostics: vscode.Diagnostic[] = [];
+	if (enableDiagnostics && document) {
         for (let lineIndex = 0; lineIndex < document.lineCount; lineIndex++) {
             const lineOfText = document.lineAt(lineIndex);
             lineChecks.forEach((x)=>{
@@ -24,6 +31,13 @@ export function updateDiagnostics(document: vscode.TextDocument, collection: vsc
                 }
             });
         }
+	}
+    return diagnostics;
+}
+
+export function updateDiagnostics(document: vscode.TextDocument, collection: vscode.DiagnosticCollection): void {
+	if (document) {
+        const diagnostics = checkDocument(document);
         collection.set(document.uri, diagnostics);
 	} else {
 		collection.clear();